+++ frontend/src/config.ts
@@ -4,7 +4,7 @@ export const stage: Stage = process.env.REACT_APP_STAGE === 'prod' ? 'prod' : 'd
 
 export const THEME = { primary: '#2b669a', colorOnPrimary: 'white' };
 
-// Determines whether stubs or API endpoints are used
+// Determines whether stubs or API endpoints are used.
 export const USE_STUBS = false;
 
 const LOCAL_DEV_ENDPOINT = 'http://localhost:8080/unidisk_war/rest/';
