+++ frontend/src/data/services/Authentication.ts
@@ -1,5 +1,14 @@
 import { User } from '../user/User';
-import { AuthProvider } from '../../remote/util';
+
+export type AuthToken = {
+  token: string;
+  userId: string;
+};
+
+export interface AuthProvider {
+  getAuthToken(): Promise<AuthToken>;
+  onTokenChanged(callback: (token: string) => void): void;
+}
 
 export default interface IAuthenticationService extends AuthProvider {
   login(email: string, password: string): Promise<User | null>;
